Try to import `rdkafka-2.10` from `nixos-25.05` and everything else from `nixos-24.11`
libunwind is broken in "nixpkgs#pkgsCross.musl64.pkgsStatic.pkgsLLVM.libunwind", see https://github.com/NixOS/nixpkgs/pull/417354

This hack is not the best solution, but it's the only one I was able to hack together. I think it would be better to override `libunwind` in `stdenv`, but I wasn't able to do that.